Continue error system improvements
This commit continues the program described in the previous mainline commit.
Here, we change the signature of the recoverable error hook and as a result
enforce that the hook is called at most once. This uncovered a few bugs in
the code. Additionally, I changed many of the fatal errors to be recoverable
errors.