texlang: add a VM execution stack
Texlang errors contain information about the execution
stack at the point of the error - e.g., which TeX commands were currently
running, which TeX expansions were happening, etc. Though TeX is not a standard
programming language it still has a notion of a stack like this.

The stack information (aka propagation information) was previously built using
the _Rust_ stack: as the error bubbled up the Rust call stack corresponding
to the TeX stack, the stack information was added. The problem is that this
doesn't play well with the recoverable error hook because the hook sees the
error before its propagated. In particular, if the hook just logs the error
and moves on, any error logging won't include the stack information.

With this commit the VM instead maintains an explicit execution stack. When
an error occurs the VM can provide the stack trace immediately. Potentially
in the future TeX debugging tools could be built using this stack.