texlang-stdlib: handler missing control sequence
Commands like \def and \countdef are supposed to be followed by a control
sequence to be defined. If it is missing, TeX continues processing the
definition but throws away the result. This commit handles this case for
commands in Texlang stdlib.

In Knuth's TeX, the command assignment is thrown away by assigining the result
to a special "frozen cs" slot in the eqtb. I have read the source code and
I believe it is not possible to observe this assignment in a TeX document.
Thus, this commit just skips perorming the assignment if the control sequence
is missing.