Font loading subsystem
Adds support for loading fonts via the \font command. Currently only loading
.tfm files is supported.

The main new abstraction in this commit is a "font repo" which is a place
where fonts are stored after they are loaded. The motivation for this
abstraction is to decouple the implementation of \font (which loads the font)
from a specific typesetting engine (which uses the fonts). The engine will
probably implement font repo and thus \font can load fonts into it,
irrespective of other implementation details of the engine.

Also implements \fontname and font type variables.