Fix bug in optimized \expandafter
The optimized \expandafter repeatedly expanded the input until the first token
was not an \expandafter token. But this is incorrect: expansion commands are
only supposed to expand once and then return. In future we may add code to
the VM to detect whether we should be expanding fully, and then we can re-add
the optimization or a similar optimization.