Make all end-of-input errors fatal
In my recent work on recoverable errors for \def, I realized that making the
end of input cases recoverable basically duplicated the fatal error case. The
reason I realized is that end of input errors aren't truly recoverable, because
the end of input condition is permanent. If you recover from an end of input
error you're just going to hit it again when the next piece of code tries
to read from the stream.

With the commit, we make these kinds of errors always fatal. With this change
the signature of TokenStream::next can be changed to always return a token
or an error, which simplifies some of the callsites, though an error type
now has to be provided.

By the way, Knuth's TeX doesn't handle end of input errors gracefully at all.
They are always fatal, and the error message you get is independent of where
the condition happened. This would actually be easier to replicate than what
we're doing here...but for the moment we try to have more contextual errors.