Decouple scaled and fix word number
The code was based on the incorrect assumption that
scaled numbers (in TeX) and fix word numbers (in TFM)
were the same. They are not: scaled numbers are fixed
point numbers with 16 binary digits after the point,
whereas fix word have 20 binary digits after the point.
Fix words are converted to scaled numbers are part of
reading the TFM file.

This commit:
- Renames the font crate to the core crate. Now that
  Number is coming from a font format, the crate isn't
  font specific.
- Renames core::Number to core::Scaled
- Copies core::Scaled into tfm::FixWord and updates
  references accordingly.
- Fixes the single place where core::Scaled
  is assumed to have 20 binary fractional digits.