tfm: upgrade version of ariadne
I'm using ariadne in a new crate and realized the version here is a little
old and uses some older APIs. So this commit updates the version so there
is a uniform version in the project.

I also remove some ariadne helper code which is actually good code because
it gets around some problems with "default" usage of ariadne. However it's
probably not worth the hassle of maintaining because these problems only
happen on the error path.