texlang: improve Parsable trait
Rather than a generic trait with a single non-generic method, make it a
non-generic trait with a single generic method. This makes is possible to
specify the trait point in places where the state type is not present. This
is also the "right" way to do it, I think: I don't think this was a use case
for generic traits.