tfm: remove Font type
My initial design for the tfm crate had 3 main types: a type representing
a TFM file, a type representing a PL file, and then another type representing
the data in a compiled form that would be used inside the typesetting engine.
This last type would be "optimized" to make typesetting fast.

Recently I started working on the Knuth-Plass algorithm and I realized that
what makes a type "optimal" is based on how it is used. E.g., in Knuth-Plass
only the widths of the characters are read. Thus due to cache efficiency it's
best to pack all the widths together. For this reason I now think that having
an optimized type in the TFM crate doesn't really make sense - instead, the
Boxworks engine will read TFM data and then store it in a way that's
optimal for Boxworks.

This commit thus removes the "optimal" Font type.