texlang: yet more errors system work
This is the latest commit that changes the Texlang errors system! Although
it's not as radical as previous ones and is more of an incremental improvement.

The commit changes the error payload of the return type of Texlang functions
from Box of an error type to an opaque "stop signal". This change means the
code more accurately reflects the semantics of error propagation in Texlang,
in which errors are never supposed to be handled. I additionally add a check
and panic that stop signals are not ignored.