box tex hlist: support non-cmr10 fonts
This is being added to facilatate lig/kern testing. My plan is to create
a very custom lig/kern program (in ligaroo.pl) that exercises the different
cases. With this box command change, I can verify that output of the lig/kern
command in Knuth's TeX, and then write unit tests for my implementation.